[level 1] Title: 최대공약수와 최소공배수, Time: 0.02 ms, Memory: 67.1 MB -BaekjoonHub

+++ 프로그래머스/1/12940. 최대공약수와 최소공배수/최대공약수와 최소공배수.java
@@ -1,16 +1,16 @@
 import java.util.*;
 
 class Solution {
-    public int[] solution(int n, int m) {
+    public long[] solution(int n, int m) {
         
         // 정답을 담을 배열
-        int[] answer = new int[2];
+        long[] answer = new long[2];
         
         // 두 수의 최대공약수
         int gcdVal = gcd(n, m);
         
         // 두 수의 최소공배수
-        int lcmVal = (n * m) / gcdVal; 
+        long lcmVal = ((long)n * m) / gcdVal; 
         
         answer[0] = gcdVal;
         answer[1] = lcmVal;
